Splits out ebook generation logic into a seperate module, in anticipation of maybe supporting multiple output formats.
